+++ app/domains/narou.ts
@@ -1,3 +1,32 @@
+export type NarouNovel = {
+  ncode: string
+  title: string
+  totalPage: number
+  currentPage: number
+  lastPublishedAt: string
+}
+
+type RegexMatchGroup = {
+  ncode: string
+  page?: string
+}
+
+type NarouPageInfo = {
+  ncode: string
+  page: number
+}
+
+export const parseNcodeAndPageFromUrlPath = (url: string): NarouPageInfo => {
+  const regex = /\/(?<ncode>n[a-zA-Z0-9]+)\/?(?<page>\d*).*$/
+  const matches = regex.exec(url)
+  if (!matches) {
+    console.error(`invalid url for narou novel : url=${url}`)
+    throw new Error(`invalid url for narou novel : url=${url}`)
+  }
+  const { ncode, page } = matches.groups as RegexMatchGroup
+  return { ncode, page: Number(page) || 0 }
+}
+
 /**
  * YYYY-MM-DD HH:mm:ss → YYYY年MM月DD日HH時mm分
  */
@@ -12,9 +41,11 @@ export const getPageInfo = (currentPage: number, totalPage: number) => {
   return `${currentPage}話／全${totalPage}話`
 }
 
+/**
+ * page=0のケースでpageの省略を考慮すべきかも
+ * ncodeのみのURLが表示用にほしい場合もあるので分けてもいいかも
+ */
 export const getOriginalNarouUrl = (ncode: string, page: number) => {
-  // page=0のケースでpageの省略を考慮すべきかも
-  // ncodeのみのURLが表示用にほしい場合もあるので分けてもいいかも
   return `https://ncode.syosetu.com/${ncode}/${page}`
 }
